fix: harden broker account projection
