--- crates/ironclaw_reborn_composition/src/product_auth_durable/tests.rs
@@ -1,2580 +1,0 @@
-use std::sync::Arc;
-
-use chrono::{Duration, Utc};
-use ironclaw_filesystem::{InMemoryBackend, ScopedFilesystem};
-use ironclaw_host_api::{
-    InvocationId, MountAlias, MountGrant, MountPermissions, SecretHandle, UserId, VirtualPath,
-};
-use ironclaw_secrets::{InMemorySecretStore, SecretStore};
-use secrecy::SecretString;
-use tokio::task::JoinSet;
-
-use super::*;
-use ironclaw_auth::{
-    AuthChallenge, AuthContinuationRef, AuthFlowKind, AuthFlowManager, AuthFlowStatus,
-    AuthInteractionService, AuthProductError, AuthProductScope, AuthProviderId, AuthSurface,
-    AuthorizationCodeHash, CredentialAccountChoiceRequest, CredentialAccountLabel,
-    CredentialAccountListRequest, CredentialAccountLookupRequest,
-    CredentialAccountSelectionRequest, CredentialAccountService, CredentialAccountStatus,
-    CredentialOwnership, ManualTokenSetupRequest, NewAuthFlow, NewCredentialAccount,
-    OAuthAuthorizationUrl, OAuthCallbackClaimRequest, OAuthCallbackInput, OAuthProviderExchange,
-    OpaqueStateHash, PkceVerifierHash, ProviderScope, SecretSubmitRequest,
-};
-
-fn test_scope() -> AuthProductScope {
-    let resource =
-        ResourceScope::local_default(UserId::new("alice").unwrap(), InvocationId::new()).unwrap();
-    AuthProductScope::new(resource, AuthSurface::Web)
-}
-
-fn test_filesystem() -> Arc<ScopedFilesystem<InMemoryBackend>> {
-    let mounts = ironclaw_host_api::MountView::new(vec![MountGrant::new(
-        MountAlias::new("/secrets").unwrap(),
-        VirtualPath::new("/tenants/test/users/alice/secrets").unwrap(),
-        MountPermissions::read_write_list_delete(),
-    )])
-    .unwrap();
-    Arc::new(ScopedFilesystem::with_fixed_view(
-        Arc::new(InMemoryBackend::new()),
-        mounts,
-    ))
-}
-
-fn test_service(
-    filesystem: Arc<ScopedFilesystem<InMemoryBackend>>,
-    secret_store: Arc<dyn SecretStore>,
-) -> FilesystemAuthProductServices<InMemoryBackend> {
-    FilesystemAuthProductServices::new(
-        filesystem,
-        secret_store,
-        Arc::new(super::NoopBrokerAccountProjector),
-    )
-}
-
-/// Variant of [`test_service`] that installs a recording projector and
-/// returns it so the caller can assert projection invocations after
-/// driving a flow.  Used by `broker_projection_tests` below.
-fn test_service_with_recording_projector(
-    filesystem: Arc<ScopedFilesystem<InMemoryBackend>>,
-    secret_store: Arc<dyn SecretStore>,
-) -> (
-    FilesystemAuthProductServices<InMemoryBackend>,
-    Arc<super::broker_projection::test_support::RecordingBrokerProjector>,
-) {
-    let projector =
-        Arc::new(super::broker_projection::test_support::RecordingBrokerProjector::new());
-    let service = FilesystemAuthProductServices::new(
-        filesystem,
-        secret_store,
-        Arc::clone(&projector) as Arc<dyn super::broker_projection::BrokerAccountProjector>,
-    );
-    (service, projector)
-}
-
-fn google_provider() -> AuthProviderId {
-    AuthProviderId::new("google").unwrap()
-}
-
-fn account_label() -> CredentialAccountLabel {
-    CredentialAccountLabel::new("Alice Google").unwrap()
-}
-
-fn fake_digest(value: &str) -> String {
-    format!(
-        "{:064x}",
-        value.bytes().fold(0_u64, |hash, byte| {
-            hash.wrapping_mul(31).wrapping_add(u64::from(byte))
-        })
-    )
-}
-
-fn state_hash(value: &str) -> OpaqueStateHash {
-    OpaqueStateHash::new(fake_digest(value)).unwrap()
-}
-
-fn pkce_hash(value: &str) -> PkceVerifierHash {
-    PkceVerifierHash::new(fake_digest(value)).unwrap()
-}
-
-fn code_hash(value: &str) -> AuthorizationCodeHash {
-    AuthorizationCodeHash::new(fake_digest(value)).unwrap()
-}
-
-#[tokio::test]
-async fn filesystem_accounts_survive_service_recreation() {
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = test_service(Arc::clone(&filesystem), Arc::clone(&secret_store));
-
-    let created = service
-        .create_account(NewCredentialAccount {
-            scope: scope.clone(),
-            provider: google_provider(),
-            label: account_label(),
-            status: CredentialAccountStatus::Configured,
-            ownership: CredentialOwnership::UserReusable,
-            owner_extension: None,
-            granted_extensions: Vec::new(),
-            access_secret: Some(SecretHandle::new("google-access").unwrap()),
-            refresh_secret: Some(SecretHandle::new("google-refresh").unwrap()),
-            scopes: vec![ProviderScope::new("gmail.readonly").unwrap()],
-        })
-        .await
-        .unwrap();
-
-    let recreated = test_service(Arc::clone(&filesystem), secret_store);
-    let loaded = recreated
-        .get_account(CredentialAccountLookupRequest::new(
-            scope.clone(),
-            created.id,
-        ))
-        .await
-        .unwrap()
-        .expect("account should be durable");
-    assert_eq!(loaded.id, created.id);
-    assert_eq!(loaded.access_secret, created.access_secret);
-
-    let page = recreated
-        .list_accounts(CredentialAccountListRequest::new(scope, google_provider()))
-        .await
-        .unwrap();
-    assert_eq!(page.accounts.len(), 1);
-    assert_eq!(page.accounts[0].id, created.id);
-}
-
-#[tokio::test]
-async fn filesystem_manual_token_submit_stores_secret_and_dedupes_replay() {
-    let filesystem = test_filesystem();
-    let concrete_secret_store = Arc::new(InMemorySecretStore::new());
-    let secret_store: Arc<dyn SecretStore> = concrete_secret_store.clone();
-    let scope = test_scope();
-    let service = test_service(filesystem, secret_store);
-
-    let challenge = service
-        .request_secret_input(ManualTokenSetupRequest {
-            scope: scope.clone(),
-            provider: google_provider(),
-            label: account_label(),
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: None,
-            expires_at: Utc::now() + Duration::minutes(5),
-        })
-        .await
-        .unwrap();
-    let AuthChallenge::ManualTokenRequired { interaction_id, .. } = challenge else {
-        panic!("expected manual token challenge");
-    };
-
-    let result = service
-        .submit_manual_token(
-            &scope,
-            SecretSubmitRequest {
-                interaction_id,
-                secret: SecretString::from("manual-token-value"),
-            },
-        )
-        .await
-        .unwrap();
-    assert_eq!(result.status, CredentialAccountStatus::Configured);
-
-    let account = service
-        .get_account(CredentialAccountLookupRequest::new(
-            scope.clone(),
-            result.account_id,
-        ))
-        .await
-        .unwrap()
-        .expect("manual token submit should create account");
-    let access_secret = account.access_secret.expect("manual token secret handle");
-    assert!(
-        concrete_secret_store
-            .metadata(&scope.resource, &access_secret)
-            .await
-            .unwrap()
-            .is_some()
-    );
-
-    let replay = service
-        .submit_manual_token(
-            &scope,
-            SecretSubmitRequest {
-                interaction_id,
-                secret: SecretString::from("manual-token-value"),
-            },
-        )
-        .await
-        .expect_err("manual token submit should be one-shot");
-    assert_eq!(replay, AuthProductError::UnknownOrExpiredFlow);
-}
-
-#[tokio::test]
-async fn filesystem_oauth_callback_claim_is_one_shot_and_completion_persists() {
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = test_service(Arc::clone(&filesystem), Arc::clone(&secret_store));
-
-    let flow = service
-        .create_flow(NewAuthFlow {
-            scope: scope.clone(),
-            kind: AuthFlowKind::IntegrationCredential,
-            provider: google_provider(),
-            challenge: AuthChallenge::OAuthUrl {
-                authorization_url: OAuthAuthorizationUrl::new("https://provider.example/oauth")
-                    .unwrap(),
-                expires_at: Utc::now() + Duration::minutes(5),
-            },
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: None,
-            opaque_state_hash: Some(state_hash("state")),
-            pkce_verifier_hash: Some(pkce_hash("pkce")),
-            expires_at: Utc::now() + Duration::minutes(5),
-        })
-        .await
-        .unwrap();
-    let claim = OAuthCallbackClaimRequest {
-        flow_id: flow.id,
-        opaque_state_hash: state_hash("state"),
-        provider: google_provider(),
-        pkce_verifier_hash: pkce_hash("pkce"),
-    };
-
-    let claimed = service
-        .claim_oauth_callback(&scope, claim.clone())
-        .await
-        .unwrap();
-    assert_eq!(claimed.status, AuthFlowStatus::CallbackReceived);
-
-    let second_claim = service
-        .claim_oauth_callback(&scope, claim.clone())
-        .await
-        .expect_err("in-flight callback claim must be one-shot");
-    assert_eq!(second_claim, AuthProductError::FlowAlreadyTerminal);
-
-    let completed = service
-        .complete_oauth_callback(
-            &scope,
-            OAuthCallbackInput {
-                flow_id: flow.id,
-                opaque_state_hash: state_hash("state"),
-                outcome: ironclaw_auth::ProviderCallbackOutcome::Authorized {
-                    exchange: OAuthProviderExchange {
-                        provider: google_provider(),
-                        account_label: account_label(),
-                        authorization_code_hash: code_hash("code"),
-                        pkce_verifier_hash: pkce_hash("pkce"),
-                        access_secret: SecretHandle::new("oauth-access").unwrap(),
-                        refresh_secret: Some(SecretHandle::new("oauth-refresh").unwrap()),
-                        scopes: vec![ProviderScope::new("gmail.readonly").unwrap()],
-                        account_id: None,
-                    },
-                },
-            },
-        )
-        .await
-        .unwrap();
-    assert_eq!(completed.status, AuthFlowStatus::Completed);
-    assert!(completed.credential_account_id.is_some());
-
-    let emitted_at = Utc::now();
-    service
-        .mark_continuation_dispatched(&scope, flow.id, emitted_at)
-        .await
-        .unwrap();
-
-    let recreated = test_service(Arc::clone(&filesystem), secret_store);
-    let stored = recreated
-        .get_flow(&scope, flow.id)
-        .await
-        .unwrap()
-        .expect("completed flow should be durable");
-    assert_eq!(stored.status, AuthFlowStatus::Completed);
-    assert_eq!(stored.continuation_emitted_at, Some(emitted_at));
-
-    let completed_replay = recreated
-        .claim_oauth_callback(&scope, claim)
-        .await
-        .expect("completed callback replay should not reclaim provider exchange");
-    assert_eq!(completed_replay.status, AuthFlowStatus::Completed);
-    assert_eq!(completed_replay.continuation_emitted_at, Some(emitted_at));
-}
-
-#[tokio::test]
-async fn filesystem_manual_token_submit_allows_only_one_concurrent_consumer() {
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = Arc::new(test_service(filesystem, secret_store));
-
-    let challenge = service
-        .request_secret_input(ManualTokenSetupRequest {
-            scope: scope.clone(),
-            provider: google_provider(),
-            label: account_label(),
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: None,
-            expires_at: Utc::now() + Duration::minutes(5),
-        })
-        .await
-        .unwrap();
-    let AuthChallenge::ManualTokenRequired { interaction_id, .. } = challenge else {
-        panic!("expected manual token challenge");
-    };
-
-    let mut tasks = JoinSet::new();
-    for value in ["first-token", "second-token"] {
-        let service = Arc::clone(&service);
-        let scope = scope.clone();
-        tasks.spawn(async move {
-            service
-                .submit_manual_token(
-                    &scope,
-                    SecretSubmitRequest {
-                        interaction_id,
-                        secret: SecretString::from(value),
-                    },
-                )
-                .await
-        });
-    }
-
-    let mut successes = 0;
-    let mut consumed_rejections = 0;
-    while let Some(result) = tasks.join_next().await {
-        match result.unwrap() {
-            Ok(_) => successes += 1,
-            Err(AuthProductError::UnknownOrExpiredFlow) => consumed_rejections += 1,
-            Err(error) => panic!("unexpected submit error: {error:?}"),
-        }
-    }
-
-    assert_eq!(successes, 1);
-    assert_eq!(consumed_rejections, 1);
-}
-
-// ─── fix: fs_error maps VersionMismatch to BackendConflict ───────────────────
-
-#[test]
-fn fs_error_maps_version_mismatch_to_backend_conflict() {
-    use super::paths::fs_error;
-    use ironclaw_filesystem::{FilesystemError, FilesystemOperation};
-    use ironclaw_host_api::VirtualPath;
-
-    let version_mismatch = FilesystemError::VersionMismatch {
-        path: VirtualPath::new("/secrets/test").unwrap(),
-        expected: None,
-        found: None,
-    };
-    assert_eq!(
-        fs_error(version_mismatch),
-        AuthProductError::BackendConflict,
-        "VersionMismatch must map to BackendConflict, not BackendUnavailable"
-    );
-
-    let backend_err = FilesystemError::Backend {
-        path: VirtualPath::new("/secrets/test").unwrap(),
-        operation: FilesystemOperation::ReadFile,
-        reason: "io error".to_string(),
-    };
-    assert_eq!(
-        fs_error(backend_err),
-        AuthProductError::BackendUnavailable,
-        "non-CAS errors must still map to BackendUnavailable"
-    );
-}
-
-// ─── fix: mark_continuation_dispatched is idempotent ─────────────────────────
-
-#[tokio::test]
-async fn filesystem_oauth_continuation_marker_is_idempotent() {
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = test_service(Arc::clone(&filesystem), Arc::clone(&secret_store));
-
-    let flow = service
-        .create_flow(NewAuthFlow {
-            scope: scope.clone(),
-            kind: AuthFlowKind::IntegrationCredential,
-            provider: google_provider(),
-            challenge: AuthChallenge::OAuthUrl {
-                authorization_url: OAuthAuthorizationUrl::new("https://provider.example/oauth")
-                    .unwrap(),
-                expires_at: Utc::now() + Duration::minutes(5),
-            },
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: None,
-            opaque_state_hash: Some(state_hash("s")),
-            pkce_verifier_hash: Some(pkce_hash("p")),
-            expires_at: Utc::now() + Duration::minutes(5),
-        })
-        .await
-        .unwrap();
-
-    // Complete the flow so mark_continuation_dispatched is valid.
-    service
-        .claim_oauth_callback(
-            &scope,
-            OAuthCallbackClaimRequest {
-                flow_id: flow.id,
-                opaque_state_hash: state_hash("s"),
-                provider: google_provider(),
-                pkce_verifier_hash: pkce_hash("p"),
-            },
-        )
-        .await
-        .unwrap();
-    service
-        .complete_oauth_callback(
-            &scope,
-            OAuthCallbackInput {
-                flow_id: flow.id,
-                opaque_state_hash: state_hash("s"),
-                outcome: ironclaw_auth::ProviderCallbackOutcome::Authorized {
-                    exchange: OAuthProviderExchange {
-                        provider: google_provider(),
-                        account_label: account_label(),
-                        authorization_code_hash: code_hash("c"),
-                        pkce_verifier_hash: pkce_hash("p"),
-                        access_secret: SecretHandle::new("access").unwrap(),
-                        refresh_secret: None,
-                        scopes: vec![],
-                        account_id: None,
-                    },
-                },
-            },
-        )
-        .await
-        .unwrap();
-
-    let first_at = Utc::now();
-    let first = service
-        .mark_continuation_dispatched(&scope, flow.id, first_at)
-        .await
-        .unwrap();
-    assert_eq!(first.continuation_emitted_at, Some(first_at));
-
-    // Second call with a different timestamp must NOT overwrite.
-    let second_at = first_at + Duration::seconds(1);
-    let second = service
-        .mark_continuation_dispatched(&scope, flow.id, second_at)
-        .await
-        .unwrap();
-    assert_eq!(
-        second.continuation_emitted_at,
-        Some(first_at),
-        "idempotent: second call must not overwrite the first emitted_at"
-    );
-}
-
-// ─── fix: manual-token submit cleans up secret on write failure ───────────────
-
-#[tokio::test]
-async fn filesystem_manual_token_rotation_removes_previous_secret() {
-    // Tests the update_binding path in create_or_update_manual_token_account:
-    // after a successful token rotation the OLD access secret must be purged
-    // from SecretStore so it does not accumulate orphaned material.
-    use ironclaw_auth::{
-        CredentialAccountUpdateBinding, ManualTokenSetupRequest, SecretSubmitRequest,
-    };
-
-    let filesystem = test_filesystem();
-    let concrete_secret_store = Arc::new(InMemorySecretStore::new());
-    let secret_store: Arc<dyn SecretStore> = concrete_secret_store.clone();
-    let scope = test_scope();
-    let service = test_service(Arc::clone(&filesystem), Arc::clone(&secret_store));
-
-    // --- First submit: create the account via the no-binding path. ---
-    let challenge1 = service
-        .request_secret_input(ManualTokenSetupRequest {
-            scope: scope.clone(),
-            provider: google_provider(),
-            label: account_label(),
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: None,
-            expires_at: Utc::now() + Duration::minutes(5),
-        })
-        .await
-        .unwrap();
-    let AuthChallenge::ManualTokenRequired {
-        interaction_id: iid1,
-        ..
-    } = challenge1
-    else {
-        panic!("expected ManualTokenRequired");
-    };
-    let result1 = service
-        .submit_manual_token(
-            &scope,
-            SecretSubmitRequest {
-                interaction_id: iid1,
-                secret: SecretString::from("token-v1"),
-            },
-        )
-        .await
-        .unwrap();
-    let account_id = result1.account_id;
-
-    // Grab the first-generation secret handle.
-    let account_after_v1 = service
-        .get_account(ironclaw_auth::CredentialAccountLookupRequest::new(
-            scope.clone(),
-            account_id,
-        ))
-        .await
-        .unwrap()
-        .unwrap();
-    let old_handle = account_after_v1
-        .access_secret
-        .clone()
-        .expect("v1 access_secret");
-    assert!(
-        concrete_secret_store
-            .metadata(&scope.resource, &old_handle)
-            .await
-            .unwrap()
-            .is_some(),
-        "v1 secret must exist in store"
-    );
-
-    // --- Second submit: rotate via update_binding to the same account. ---
-    let challenge2 = service
-        .request_secret_input(ManualTokenSetupRequest {
-            scope: scope.clone(),
-            provider: google_provider(),
-            label: account_label(),
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: Some(CredentialAccountUpdateBinding {
-                account_id,
-                ownership: CredentialOwnership::UserReusable,
-                owner_extension: None,
-                granted_extensions: vec![],
-            }),
-            expires_at: Utc::now() + Duration::minutes(5),
-        })
-        .await
-        .unwrap();
-    let AuthChallenge::ManualTokenRequired {
-        interaction_id: iid2,
-        ..
-    } = challenge2
-    else {
-        panic!("expected ManualTokenRequired for rotation");
-    };
-    service
-        .submit_manual_token(
-            &scope,
-            SecretSubmitRequest {
-                interaction_id: iid2,
-                secret: SecretString::from("token-v2"),
-            },
-        )
-        .await
-        .unwrap();
-
-    // The old handle must have been purged from SecretStore after the rotation.
-    assert!(
-        concrete_secret_store
-            .metadata(&scope.resource, &old_handle)
-            .await
-            .unwrap()
-            .is_none(),
-        "v1 secret must be purged from SecretStore after rotation"
-    );
-
-    // The new handle must be present.
-    let account_after_v2 = service
-        .get_account(ironclaw_auth::CredentialAccountLookupRequest::new(
-            scope.clone(),
-            account_id,
-        ))
-        .await
-        .unwrap()
-        .unwrap();
-    let new_handle = account_after_v2.access_secret.expect("v2 access_secret");
-    assert!(
-        concrete_secret_store
-            .metadata(&scope.resource, &new_handle)
-            .await
-            .unwrap()
-            .is_some(),
-        "v2 secret must be present in SecretStore"
-    );
-}
-
-// ─── fix: durable SecretCleanupService purges secrets on Uninstall ───────────
-
-#[tokio::test]
-async fn filesystem_cleanup_for_lifecycle_deactivates_owner_and_revokes_on_uninstall() {
-    use ironclaw_auth::{SecretCleanupAction, SecretCleanupRequest, SecretCleanupService};
-    use ironclaw_host_api::ExtensionId;
-
-    let filesystem = test_filesystem();
-    let concrete_secret_store = Arc::new(InMemorySecretStore::new());
-    let secret_store: Arc<dyn SecretStore> = concrete_secret_store.clone();
-    let scope = test_scope();
-    let service = test_service(Arc::clone(&filesystem), Arc::clone(&secret_store));
-
-    let ext_id = ExtensionId::new("test-ext").unwrap();
-    let access = SecretHandle::new("ext-access").unwrap();
-    let refresh = SecretHandle::new("ext-refresh").unwrap();
-
-    // Seed secret material.
-    use secrecy::SecretString;
-    concrete_secret_store
-        .put(
-            scope.resource.clone(),
-            access.clone(),
-            SecretString::from("access-material"),
-        )
-        .await
-        .unwrap();
-    concrete_secret_store
-        .put(
-            scope.resource.clone(),
-            refresh.clone(),
-            SecretString::from("refresh-material"),
-        )
-        .await
-        .unwrap();
-
-    // Create an extension-owned account.
-    let account = service
-        .create_account(ironclaw_auth::NewCredentialAccount {
-            scope: scope.clone(),
-            provider: google_provider(),
-            label: account_label(),
-            status: CredentialAccountStatus::Configured,
-            ownership: CredentialOwnership::ExtensionOwned,
-            owner_extension: Some(ext_id.clone()),
-            granted_extensions: vec![],
-            access_secret: Some(access.clone()),
-            refresh_secret: Some(refresh.clone()),
-            scopes: vec![],
-        })
-        .await
-        .unwrap();
-
-    // Deactivate: account should be Inactive; secrets retained.
-    let deactivate_report = service
-        .cleanup_for_lifecycle(SecretCleanupRequest {
-            scope: scope.clone(),
-            extension_id: ext_id.clone(),
-            action: SecretCleanupAction::Deactivate,
-        })
-        .await
-        .unwrap();
-    assert_eq!(deactivate_report.retained_accounts, vec![account.id]);
-    assert!(
-        concrete_secret_store
-            .metadata(&scope.resource, &access)
-            .await
-            .unwrap()
-            .is_some(),
-        "Deactivate must retain secret material"
-    );
-
-    // Uninstall: account revoked, secrets purged from SecretStore.
-    let uninstall_report = service
-        .cleanup_for_lifecycle(SecretCleanupRequest {
-            scope: scope.clone(),
-            extension_id: ext_id.clone(),
-            action: SecretCleanupAction::Uninstall,
-        })
-        .await
-        .unwrap();
-    assert_eq!(uninstall_report.revoked_accounts, vec![account.id]);
-    assert!(
-        concrete_secret_store
-            .metadata(&scope.resource, &access)
-            .await
-            .unwrap()
-            .is_none(),
-        "Uninstall must delete access secret from SecretStore"
-    );
-    assert!(
-        concrete_secret_store
-            .metadata(&scope.resource, &refresh)
-            .await
-            .unwrap()
-            .is_none(),
-        "Uninstall must delete refresh secret from SecretStore"
-    );
-}
-
-// ─── fix: lock-cache weak-reference GC actually shrinks the map ──────────────
-
-#[tokio::test]
-async fn filesystem_lock_cache_drops_weak_entries_after_release() {
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let service = test_service(filesystem, secret_store);
-
-    {
-        // Acquire a lock for key A and drop the guard immediately.
-        let lock_a = service.lock_for("account:key-a".to_string());
-        let _guard_a = lock_a.lock().await;
-        // guard_a dropped at end of this block; Arc<Mutex> dropped too after lock_a drops.
-    }
-    // After key-A's Arc dropped, the next call to lock_for should evict the
-    // dead weak reference. We trigger eviction via lock_for on a different key.
-    let _lock_b = service.lock_for("account:key-b".to_string());
-
-    // Verify key-A is gone: requesting it again must produce a *new* Arc (i.e.
-    // a fresh Mutex), not the evicted weak ref.
-    let lock_a2 = service.lock_for("account:key-a".to_string());
-    // The new lock should be unlocked (no one holds it).
-    assert!(
-        lock_a2.try_lock().is_ok(),
-        "re-acquired key-a must be unlocked"
-    );
-}
-
-// ─── fix: manual-token expiry branch ─────────────────────────────────────────
-
-#[tokio::test]
-async fn filesystem_manual_token_submit_rejects_expired_interaction() {
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = test_service(filesystem, secret_store);
-
-    // Create an interaction that is already past its expiry.
-    let challenge = service
-        .request_secret_input(ManualTokenSetupRequest {
-            scope: scope.clone(),
-            provider: google_provider(),
-            label: account_label(),
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: None,
-            // Expired immediately.
-            expires_at: Utc::now() - Duration::seconds(1),
-        })
-        .await
-        .unwrap();
-    let AuthChallenge::ManualTokenRequired { interaction_id, .. } = challenge else {
-        panic!("expected ManualTokenRequired");
-    };
-
-    let err = service
-        .submit_manual_token(
-            &scope,
-            SecretSubmitRequest {
-                interaction_id,
-                secret: SecretString::from("too-late"),
-            },
-        )
-        .await
-        .expect_err("expired interaction must be rejected");
-    assert_eq!(
-        err,
-        AuthProductError::UnknownOrExpiredFlow,
-        "expired interaction must return UnknownOrExpiredFlow"
-    );
-}
-
-// ─── UnavailableAuthProviderClient validates before returning error ───────────
-
-#[tokio::test]
-async fn unavailable_auth_provider_client_validates_before_returning_backend_unavailable() {
-    use super::provider::UnavailableAuthProviderClient;
-    use ironclaw_auth::{
-        AuthProviderClient, OAuthAuthorizationCode, OAuthProviderCallbackRequest,
-        OAuthProviderExchangeContext, OAuthProviderRefreshRequest,
-    };
-    use secrecy::SecretString;
-
-    let client = UnavailableAuthProviderClient;
-
-    let ctx = OAuthProviderExchangeContext {
-        scope: test_scope(),
-        flow_id: ironclaw_auth::AuthFlowId::new(),
-    };
-
-    // Valid request must return BackendUnavailable (no provider configured) after
-    // the internal validate_provider_callback_request guard passes.
-    let valid = OAuthProviderCallbackRequest {
-        provider: google_provider(),
-        account_label: account_label(),
-        authorization_code: OAuthAuthorizationCode::new(SecretString::from("real-code")).unwrap(),
-        authorization_code_hash: code_hash("c"),
-        pkce_verifier: ironclaw_auth::PkceVerifierSecret::new(SecretString::from("real-verifier"))
-            .unwrap(),
-        pkce_verifier_hash: pkce_hash("p"),
-        scopes: vec![],
-    };
-    let err = client.exchange_callback(ctx, valid).await.unwrap_err();
-    assert_eq!(
-        err,
-        AuthProductError::BackendUnavailable,
-        "valid request must reach BackendUnavailable (no provider configured)"
-    );
-
-    // 3. refresh_token always BackendUnavailable.
-    let refresh_err = client
-        .refresh_token(OAuthProviderRefreshRequest {
-            scope: test_scope(),
-            account_id: CredentialAccountId::new(),
-            provider: google_provider(),
-            refresh_secret: SecretHandle::new("r").unwrap(),
-            scopes: vec![],
-        })
-        .await
-        .unwrap_err();
-    assert_eq!(refresh_err, AuthProductError::BackendUnavailable);
-}
-
-// ─── validate_account_list_request boundary cases ────────────────────────────
-
-#[tokio::test]
-async fn filesystem_list_accounts_rejects_zero_and_oversized_limit() {
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = test_service(filesystem, secret_store);
-
-    // limit = 0.
-    let err = service
-        .list_accounts(
-            CredentialAccountListRequest::new(scope.clone(), google_provider()).with_limit(0),
-        )
-        .await
-        .expect_err("limit=0 must be rejected");
-    assert!(matches!(err, AuthProductError::InvalidRequest { .. }));
-
-    // limit = MAX + 1.
-    let err = service
-        .list_accounts(
-            CredentialAccountListRequest::new(scope.clone(), google_provider())
-                .with_limit(CredentialAccountListRequest::MAX_LIMIT + 1),
-        )
-        .await
-        .expect_err("limit > MAX must be rejected");
-    assert!(matches!(err, AuthProductError::InvalidRequest { .. }));
-
-    // Cursor + pagination: 2 accounts, limit=1 → next_cursor present.
-    for i in 0..2u8 {
-        service
-            .create_account(ironclaw_auth::NewCredentialAccount {
-                scope: scope.clone(),
-                provider: google_provider(),
-                label: CredentialAccountLabel::new(format!("User {i}")).unwrap(),
-                status: CredentialAccountStatus::Configured,
-                ownership: CredentialOwnership::UserReusable,
-                owner_extension: None,
-                granted_extensions: vec![],
-                access_secret: None,
-                refresh_secret: None,
-                scopes: vec![],
-            })
-            .await
-            .unwrap();
-    }
-    let page = service
-        .list_accounts(
-            CredentialAccountListRequest::new(scope.clone(), google_provider()).with_limit(1),
-        )
-        .await
-        .unwrap();
-    assert_eq!(page.accounts.len(), 1);
-    assert!(
-        page.next_cursor.is_some(),
-        "second page must have next_cursor"
-    );
-}
-
-// ─── zmanian follow-up #1: OAuth re-auth must purge previous secret handles ──
-
-#[tokio::test]
-async fn filesystem_oauth_reauth_purges_previous_provider_secrets() {
-    // After a successful OAuth re-auth (exchange.account_id == Some(_)),
-    // the OLD access and refresh secret handles must be deleted from SecretStore
-    // so repeated re-auths do not accumulate dead handles.
-    use ironclaw_auth::{CredentialAccountUpdateBinding, ProviderCallbackOutcome};
-    use ironclaw_secrets::SecretMaterial;
-
-    let filesystem = test_filesystem();
-    let concrete_secret_store = Arc::new(InMemorySecretStore::new());
-    let secret_store: Arc<dyn SecretStore> = concrete_secret_store.clone();
-    let scope = test_scope();
-    let service = test_service(Arc::clone(&filesystem), Arc::clone(&secret_store));
-
-    // ── Step 1: initial OAuth flow creates a new account ─────────────────────
-    let flow1 = service
-        .create_flow(NewAuthFlow {
-            scope: scope.clone(),
-            kind: AuthFlowKind::IntegrationCredential,
-            provider: google_provider(),
-            challenge: AuthChallenge::OAuthUrl {
-                authorization_url: OAuthAuthorizationUrl::new("https://provider.example/oauth")
-                    .unwrap(),
-                expires_at: Utc::now() + Duration::minutes(5),
-            },
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: None,
-            opaque_state_hash: Some(state_hash("state1")),
-            pkce_verifier_hash: Some(pkce_hash("pkce1")),
-            expires_at: Utc::now() + Duration::minutes(5),
-        })
-        .await
-        .unwrap();
-
-    service
-        .claim_oauth_callback(
-            &scope,
-            OAuthCallbackClaimRequest {
-                flow_id: flow1.id,
-                opaque_state_hash: state_hash("state1"),
-                provider: google_provider(),
-                pkce_verifier_hash: pkce_hash("pkce1"),
-            },
-        )
-        .await
-        .unwrap();
-
-    let access_v1 = SecretHandle::new("oauth-access-v1").unwrap();
-    let refresh_v1 = SecretHandle::new("oauth-refresh-v1").unwrap();
-    // Pre-populate SecretStore to simulate provider client having stored these
-    // handles; this lets us verify they are purged on re-auth.
-    concrete_secret_store
-        .put(
-            scope.resource.clone(),
-            access_v1.clone(),
-            SecretMaterial::from("access-token-v1"),
-        )
-        .await
-        .unwrap();
-    concrete_secret_store
-        .put(
-            scope.resource.clone(),
-            refresh_v1.clone(),
-            SecretMaterial::from("refresh-token-v1"),
-        )
-        .await
-        .unwrap();
-
-    let completed1 = service
-        .complete_oauth_callback(
-            &scope,
-            OAuthCallbackInput {
-                flow_id: flow1.id,
-                opaque_state_hash: state_hash("state1"),
-                outcome: ProviderCallbackOutcome::Authorized {
-                    exchange: OAuthProviderExchange {
-                        provider: google_provider(),
-                        account_label: account_label(),
-                        authorization_code_hash: code_hash("code1"),
-                        pkce_verifier_hash: pkce_hash("pkce1"),
-                        access_secret: access_v1.clone(),
-                        refresh_secret: Some(refresh_v1.clone()),
-                        scopes: vec![ProviderScope::new("gmail.readonly").unwrap()],
-                        account_id: None,
-                    },
-                },
-            },
-        )
-        .await
-        .unwrap();
-    let account_id = completed1
-        .credential_account_id
-        .expect("first OAuth flow must produce a credential account");
-
-    // v1 handles must be present before re-auth.
-    assert!(
-        concrete_secret_store
-            .metadata(&scope.resource, &access_v1)
-            .await
-            .unwrap()
-            .is_some(),
-        "v1 access handle must exist before re-auth"
-    );
-    assert!(
-        concrete_secret_store
-            .metadata(&scope.resource, &refresh_v1)
-            .await
-            .unwrap()
-            .is_some(),
-        "v1 refresh handle must exist before re-auth"
-    );
-
-    // ── Step 2: re-auth flow bound to the existing account ───────────────────
-    let flow2 = service
-        .create_flow(NewAuthFlow {
-            scope: scope.clone(),
-            kind: AuthFlowKind::IntegrationCredential,
-            provider: google_provider(),
-            challenge: AuthChallenge::OAuthUrl {
-                authorization_url: OAuthAuthorizationUrl::new("https://provider.example/oauth")
-                    .unwrap(),
-                expires_at: Utc::now() + Duration::minutes(5),
-            },
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: Some(CredentialAccountUpdateBinding {
-                account_id,
-                ownership: CredentialOwnership::UserReusable,
-                owner_extension: None,
-                granted_extensions: vec![],
-            }),
-            opaque_state_hash: Some(state_hash("state2")),
-            pkce_verifier_hash: Some(pkce_hash("pkce2")),
-            expires_at: Utc::now() + Duration::minutes(5),
-        })
-        .await
-        .unwrap();
-
-    service
-        .claim_oauth_callback(
-            &scope,
-            OAuthCallbackClaimRequest {
-                flow_id: flow2.id,
-                opaque_state_hash: state_hash("state2"),
-                provider: google_provider(),
-                pkce_verifier_hash: pkce_hash("pkce2"),
-            },
-        )
-        .await
-        .unwrap();
-
-    let access_v2 = SecretHandle::new("oauth-access-v2").unwrap();
-    let refresh_v2 = SecretHandle::new("oauth-refresh-v2").unwrap();
-    concrete_secret_store
-        .put(
-            scope.resource.clone(),
-            access_v2.clone(),
-            SecretMaterial::from("access-token-v2"),
-        )
-        .await
-        .unwrap();
-    concrete_secret_store
-        .put(
-            scope.resource.clone(),
-            refresh_v2.clone(),
-            SecretMaterial::from("refresh-token-v2"),
-        )
-        .await
-        .unwrap();
-
-    service
-        .complete_oauth_callback(
-            &scope,
-            OAuthCallbackInput {
-                flow_id: flow2.id,
-                opaque_state_hash: state_hash("state2"),
-                outcome: ProviderCallbackOutcome::Authorized {
-                    exchange: OAuthProviderExchange {
-                        provider: google_provider(),
-                        account_label: account_label(),
-                        authorization_code_hash: code_hash("code2"),
-                        pkce_verifier_hash: pkce_hash("pkce2"),
-                        access_secret: access_v2.clone(),
-                        refresh_secret: Some(refresh_v2.clone()),
-                        scopes: vec![ProviderScope::new("gmail.readonly").unwrap()],
-                        account_id: Some(account_id),
-                    },
-                },
-            },
-        )
-        .await
-        .unwrap();
-
-    // Old handles must have been purged from SecretStore.
-    assert!(
-        concrete_secret_store
-            .metadata(&scope.resource, &access_v1)
-            .await
-            .unwrap()
-            .is_none(),
-        "v1 access handle must be purged from SecretStore after re-auth"
-    );
-    assert!(
-        concrete_secret_store
-            .metadata(&scope.resource, &refresh_v1)
-            .await
-            .unwrap()
-            .is_none(),
-        "v1 refresh handle must be purged from SecretStore after re-auth"
-    );
-
-    // New handles must remain.
-    assert!(
-        concrete_secret_store
-            .metadata(&scope.resource, &access_v2)
-            .await
-            .unwrap()
-            .is_some(),
-        "v2 access handle must be present in SecretStore after re-auth"
-    );
-    assert!(
-        concrete_secret_store
-            .metadata(&scope.resource, &refresh_v2)
-            .await
-            .unwrap()
-            .is_some(),
-        "v2 refresh handle must be present in SecretStore after re-auth"
-    );
-}
-
-// ─── [High · tests] manual-token submit cleans up secret on account write fail
-
-#[tokio::test]
-async fn filesystem_manual_token_submit_cleans_up_secret_when_account_write_fails() {
-    // create_or_update_manual_token_account (None path) stores the secret first,
-    // then calls create_account_with_id(CasExpectation::Absent). If the write
-    // fails the newly-stored secret must be deleted from SecretStore so it does
-    // not orphan in the store.
-    //
-    // Failure injection: derive the account ID that submit_manual_token will use
-    // (CredentialAccountId::from_uuid(interaction_id.as_uuid())) and write a
-    // dummy record at that path before submitting, causing CasExpectation::Absent
-    // to return VersionMismatch → BackendConflict.
-    use ironclaw_auth::CredentialAccountId;
-    use ironclaw_filesystem::CasExpectation;
-
-    let filesystem = test_filesystem();
-    let concrete_secret_store = Arc::new(InMemorySecretStore::new());
-    let secret_store: Arc<dyn SecretStore> = concrete_secret_store.clone();
-    let scope = test_scope();
-    let service = test_service(Arc::clone(&filesystem), Arc::clone(&secret_store));
-
-    // Request an interaction so we know its ID (and can derive the account path).
-    let challenge = service
-        .request_secret_input(ManualTokenSetupRequest {
-            scope: scope.clone(),
-            provider: google_provider(),
-            label: account_label(),
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: None,
-            expires_at: Utc::now() + Duration::minutes(5),
-        })
-        .await
-        .unwrap();
-    let AuthChallenge::ManualTokenRequired { interaction_id, .. } = challenge else {
-        panic!("expected ManualTokenRequired");
-    };
-
-    // Derive the same account ID the submit path will use.
-    let account_id = CredentialAccountId::from_uuid(interaction_id.as_uuid());
-
-    // Write a dummy record at that path so create_account_with_id(Absent) fails.
-    let dummy_account = ironclaw_auth::CredentialAccount {
-        id: account_id,
-        scope: scope.clone(),
-        provider: google_provider(),
-        label: account_label(),
-        status: ironclaw_auth::CredentialAccountStatus::Configured,
-        ownership: CredentialOwnership::UserReusable,
-        owner_extension: None,
-        granted_extensions: vec![],
-        access_secret: None,
-        refresh_secret: None,
-        scopes: vec![],
-        created_at: Utc::now(),
-        updated_at: Utc::now(),
-    };
-    let path = super::paths::account_path(&scope, account_id)
-        .expect("account path derivation must succeed");
-    let json = serde_json::to_vec(&dummy_account).expect("serialization must succeed");
-    use ironclaw_filesystem::{ContentType, Entry};
-    let entry = Entry::bytes(json).with_content_type(ContentType::json());
-    filesystem
-        .put(&scope.resource, &path, entry, CasExpectation::Absent)
-        .await
-        .expect("pre-create dummy account must succeed");
-
-    // Submit the token — account write will fail; cleanup must run.
-    let result = service
-        .submit_manual_token(
-            &scope,
-            SecretSubmitRequest {
-                interaction_id,
-                secret: SecretString::from("token-value"),
-            },
-        )
-        .await;
-    assert!(result.is_err(), "submit must fail when account write fails");
-
-    // The secret stored before the failing write must have been purged.
-    let access_handle = super::paths::manual_token_secret_handle(account_id, interaction_id)
-        .expect("handle derivation must succeed");
-    assert!(
-        concrete_secret_store
-            .metadata(&scope.resource, &access_handle)
-            .await
-            .unwrap()
-            .is_none(),
-        "orphaned secret must be purged from SecretStore after failed account write"
-    );
-}
-
-// ─── fix: OAuth callback CAS-conflict re-read branch ─────────────────────────
-
-#[tokio::test]
-async fn filesystem_oauth_callback_cas_conflict_reuses_concurrent_account() {
-    // Pre-create an account with the deterministic id that complete_oauth_callback
-    // derives from flow_id (CredentialAccountId::from_uuid(flow_id.as_uuid())).
-    // This simulates a concurrent callback that already created the account.
-    // The CAS-conflict branch should re-read, validate, update, and succeed.
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = test_service(Arc::clone(&filesystem), Arc::clone(&secret_store));
-
-    let flow = service
-        .create_flow(NewAuthFlow {
-            scope: scope.clone(),
-            kind: AuthFlowKind::IntegrationCredential,
-            provider: google_provider(),
-            challenge: AuthChallenge::OAuthUrl {
-                authorization_url: OAuthAuthorizationUrl::new("https://provider.example/oauth")
-                    .unwrap(),
-                expires_at: Utc::now() + Duration::minutes(5),
-            },
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: None,
-            opaque_state_hash: Some(state_hash("s2")),
-            pkce_verifier_hash: Some(pkce_hash("p2")),
-            expires_at: Utc::now() + Duration::minutes(5),
-        })
-        .await
-        .unwrap();
-
-    // Pre-seed the account with the deterministic id.
-    let preseeded_id = CredentialAccountId::from_uuid(flow.id.as_uuid());
-    service
-        .create_account_with_id(
-            preseeded_id,
-            NewCredentialAccount {
-                scope: scope.clone(),
-                provider: google_provider(),
-                label: account_label(),
-                status: CredentialAccountStatus::Configured,
-                ownership: CredentialOwnership::UserReusable,
-                owner_extension: None,
-                granted_extensions: vec![],
-                access_secret: Some(SecretHandle::new("pre-seeded-access").unwrap()),
-                refresh_secret: None,
-                scopes: vec![],
-            },
-            CasExpectation::Absent,
-        )
-        .await
-        .unwrap();
-
-    service
-        .claim_oauth_callback(
-            &scope,
-            OAuthCallbackClaimRequest {
-                flow_id: flow.id,
-                opaque_state_hash: state_hash("s2"),
-                provider: google_provider(),
-                pkce_verifier_hash: pkce_hash("p2"),
-            },
-        )
-        .await
-        .unwrap();
-
-    let completed = service
-        .complete_oauth_callback(
-            &scope,
-            OAuthCallbackInput {
-                flow_id: flow.id,
-                opaque_state_hash: state_hash("s2"),
-                outcome: ironclaw_auth::ProviderCallbackOutcome::Authorized {
-                    exchange: OAuthProviderExchange {
-                        provider: google_provider(),
-                        account_label: account_label(),
-                        authorization_code_hash: code_hash("c2"),
-                        pkce_verifier_hash: pkce_hash("p2"),
-                        access_secret: SecretHandle::new("new-access").unwrap(),
-                        refresh_secret: Some(SecretHandle::new("new-refresh").unwrap()),
-                        scopes: vec![ProviderScope::new("gmail.readonly").unwrap()],
-                        account_id: None,
-                    },
-                },
-            },
-        )
-        .await
-        .unwrap();
-
-    assert_eq!(
-        completed.credential_account_id,
-        Some(preseeded_id),
-        "CAS-conflict branch must reuse the pre-seeded account id"
-    );
-    assert_eq!(completed.status, AuthFlowStatus::Completed);
-}
-
-// ─── fix: grant-removal on non-owner account in cleanup_for_lifecycle ─────────
-
-#[tokio::test]
-async fn filesystem_cleanup_removes_grant_from_non_owner_account() {
-    use ironclaw_auth::{SecretCleanupAction, SecretCleanupRequest, SecretCleanupService};
-    use ironclaw_host_api::ExtensionId;
-
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = test_service(filesystem, secret_store);
-
-    let ext_id = ExtensionId::new("granted-ext").unwrap();
-
-    // Create user-reusable account with a grant to ext_id (not owner).
-    let account = service
-        .create_account(NewCredentialAccount {
-            scope: scope.clone(),
-            provider: google_provider(),
-            label: account_label(),
-            status: CredentialAccountStatus::Configured,
-            ownership: CredentialOwnership::UserReusable,
-            owner_extension: None,
-            granted_extensions: vec![ext_id.clone()],
-            access_secret: None,
-            refresh_secret: None,
-            scopes: vec![],
-        })
-        .await
-        .unwrap();
-
-    let report = service
-        .cleanup_for_lifecycle(SecretCleanupRequest {
-            scope: scope.clone(),
-            extension_id: ext_id.clone(),
-            action: SecretCleanupAction::Uninstall,
-        })
-        .await
-        .unwrap();
-
-    assert_eq!(
-        report.removed_grants,
-        vec![account.id],
-        "grant must be reported removed"
-    );
-    assert!(
-        report.revoked_accounts.is_empty(),
-        "non-owner account must not be revoked"
-    );
-
-    let updated = service
-        .get_account(CredentialAccountLookupRequest::new(
-            scope.clone(),
-            account.id,
-        ))
-        .await
-        .unwrap()
-        .expect("account must still exist");
-    assert!(
-        !updated.granted_extensions.contains(&ext_id),
-        "grant must be removed from account record"
-    );
-    assert_eq!(
-        updated.status,
-        CredentialAccountStatus::Configured,
-        "status must be unchanged"
-    );
-}
-
-// ─── fix: select_unique_configured_account and select_configured_account ──────
-
-#[tokio::test]
-async fn filesystem_select_unique_configured_account_single_and_multi() {
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = test_service(filesystem, secret_store);
-
-    // No accounts — CredentialMissing.
-    let err = service
-        .select_unique_configured_account(CredentialAccountSelectionRequest::new(
-            scope.clone(),
-            google_provider(),
-        ))
-        .await
-        .expect_err("no accounts must return CredentialMissing");
-    assert_eq!(err, AuthProductError::CredentialMissing);
-
-    let a1 = service
-        .create_account(NewCredentialAccount {
-            scope: scope.clone(),
-            provider: google_provider(),
-            label: account_label(),
-            status: CredentialAccountStatus::Configured,
-            ownership: CredentialOwnership::UserReusable,
-            owner_extension: None,
-            granted_extensions: vec![],
-            access_secret: None,
-            refresh_secret: None,
-            scopes: vec![],
-        })
-        .await
-        .unwrap();
-
-    // One configured — returns it.
-    let selected = service
-        .select_unique_configured_account(CredentialAccountSelectionRequest::new(
-            scope.clone(),
-            google_provider(),
-        ))
-        .await
-        .unwrap();
-    assert_eq!(selected.id, a1.id);
-
-    // Second configured — AccountSelectionRequired.
-    service
-        .create_account(NewCredentialAccount {
-            scope: scope.clone(),
-            provider: google_provider(),
-            label: CredentialAccountLabel::new("Alice Google 2").unwrap(),
-            status: CredentialAccountStatus::Configured,
-            ownership: CredentialOwnership::UserReusable,
-            owner_extension: None,
-            granted_extensions: vec![],
-            access_secret: None,
-            refresh_secret: None,
-            scopes: vec![],
-        })
-        .await
-        .unwrap();
-
-    let err = service
-        .select_unique_configured_account(CredentialAccountSelectionRequest::new(
-            scope.clone(),
-            google_provider(),
-        ))
-        .await
-        .expect_err("two configured must require selection");
-    assert_eq!(err, AuthProductError::AccountSelectionRequired);
-}
-
-#[tokio::test]
-async fn filesystem_select_configured_account_validates_provider_and_rejects_missing() {
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = test_service(filesystem, secret_store);
-
-    let account = service
-        .create_account(NewCredentialAccount {
-            scope: scope.clone(),
-            provider: google_provider(),
-            label: account_label(),
-            status: CredentialAccountStatus::Configured,
-            ownership: CredentialOwnership::UserReusable,
-            owner_extension: None,
-            granted_extensions: vec![],
-            access_secret: None,
-            refresh_secret: None,
-            scopes: vec![],
-        })
-        .await
-        .unwrap();
-
-    // Happy path.
-    let selected = service
-        .select_configured_account(CredentialAccountChoiceRequest::new(
-            scope.clone(),
-            google_provider(),
-            account.id,
-        ))
-        .await
-        .unwrap();
-    assert_eq!(selected.id, account.id);
-
-    // Non-existent account.
-    let err = service
-        .select_configured_account(CredentialAccountChoiceRequest::new(
-            scope.clone(),
-            google_provider(),
-            CredentialAccountId::new(),
-        ))
-        .await
-        .expect_err("missing account must return CredentialMissing");
-    assert_eq!(err, AuthProductError::CredentialMissing);
-
-    // Wrong provider.
-    let err = service
-        .select_configured_account(CredentialAccountChoiceRequest::new(
-            scope.clone(),
-            AuthProviderId::new("github").unwrap(),
-            account.id,
-        ))
-        .await
-        .expect_err("wrong provider must return CrossScopeDenied");
-    assert_eq!(err, AuthProductError::CrossScopeDenied);
-}
-
-// ─── tests: cancel_flow, fail_oauth_callback, complete_credential_selection ───
-
-#[tokio::test]
-async fn filesystem_cancel_flow_and_terminal_state_rejection() {
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = test_service(filesystem, secret_store);
-
-    let flow = service
-        .create_flow(NewAuthFlow {
-            scope: scope.clone(),
-            kind: AuthFlowKind::IntegrationCredential,
-            provider: google_provider(),
-            challenge: AuthChallenge::OAuthUrl {
-                authorization_url: OAuthAuthorizationUrl::new("https://provider.example/oauth")
-                    .unwrap(),
-                expires_at: Utc::now() + Duration::minutes(5),
-            },
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: None,
-            opaque_state_hash: Some(state_hash("cancel-s")),
-            pkce_verifier_hash: Some(pkce_hash("cancel-p")),
-            expires_at: Utc::now() + Duration::minutes(5),
-        })
-        .await
-        .unwrap();
-
-    let cancelled = service.cancel_flow(&scope, flow.id).await.unwrap();
-    assert_eq!(cancelled.status, AuthFlowStatus::Canceled);
-
-    // Second cancel on already-terminal flow returns Canceled error.
-    let err = service
-        .cancel_flow(&scope, flow.id)
-        .await
-        .expect_err("second cancel must fail");
-    assert_eq!(err, AuthProductError::Canceled);
-}
-
-#[tokio::test]
-async fn filesystem_fail_oauth_callback_marks_flow_failed() {
-    use ironclaw_auth::{AuthErrorCode, OAuthCallbackFailureInput};
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = test_service(filesystem, secret_store);
-
-    let flow = service
-        .create_flow(NewAuthFlow {
-            scope: scope.clone(),
-            kind: AuthFlowKind::IntegrationCredential,
-            provider: google_provider(),
-            challenge: AuthChallenge::OAuthUrl {
-                authorization_url: OAuthAuthorizationUrl::new("https://provider.example/oauth")
-                    .unwrap(),
-                expires_at: Utc::now() + Duration::minutes(5),
-            },
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: None,
-            opaque_state_hash: Some(state_hash("fail-s")),
-            pkce_verifier_hash: Some(pkce_hash("fail-p")),
-            expires_at: Utc::now() + Duration::minutes(5),
-        })
-        .await
-        .unwrap();
-
-    service
-        .claim_oauth_callback(
-            &scope,
-            OAuthCallbackClaimRequest {
-                flow_id: flow.id,
-                opaque_state_hash: state_hash("fail-s"),
-                provider: google_provider(),
-                pkce_verifier_hash: pkce_hash("fail-p"),
-            },
-        )
-        .await
-        .unwrap();
-
-    let failed = service
-        .fail_oauth_callback(
-            &scope,
-            OAuthCallbackFailureInput {
-                flow_id: flow.id,
-                opaque_state_hash: state_hash("fail-s"),
-                error: AuthErrorCode::ProviderDenied,
-            },
-        )
-        .await
-        .unwrap();
-    assert_eq!(failed.status, AuthFlowStatus::Failed);
-    assert_eq!(failed.error, Some(AuthErrorCode::ProviderDenied));
-}
-
-#[tokio::test]
-async fn filesystem_complete_credential_selection_completes_flow() {
-    use ironclaw_auth::{AuthFlowKind, CredentialSelectionInput};
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = test_service(filesystem, secret_store);
-
-    let account = service
-        .create_account(NewCredentialAccount {
-            scope: scope.clone(),
-            provider: google_provider(),
-            label: account_label(),
-            status: CredentialAccountStatus::Configured,
-            ownership: CredentialOwnership::UserReusable,
-            owner_extension: None,
-            granted_extensions: vec![],
-            access_secret: None,
-            refresh_secret: None,
-            scopes: vec![],
-        })
-        .await
-        .unwrap();
-
-    let flow = service
-        .create_flow(NewAuthFlow {
-            scope: scope.clone(),
-            kind: AuthFlowKind::IntegrationCredential,
-            provider: google_provider(),
-            challenge: AuthChallenge::AccountSelectionRequired {
-                provider: google_provider(),
-                accounts: vec![account.projection()],
-            },
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: None,
-            opaque_state_hash: None,
-            pkce_verifier_hash: None,
-            expires_at: Utc::now() + Duration::minutes(5),
-        })
-        .await
-        .unwrap();
-
-    let completed = service
-        .complete_credential_selection(
-            &scope,
-            CredentialSelectionInput {
-                flow_id: flow.id,
-                credential_account_id: account.id,
-            },
-        )
-        .await
-        .unwrap();
-    assert_eq!(completed.status, AuthFlowStatus::Completed);
-    assert_eq!(completed.credential_account_id, Some(account.id));
-}
-
-// ─── tests: create_flow update_binding validation ─────────────────────────────
-
-#[tokio::test]
-async fn filesystem_create_flow_rejects_invalid_update_binding() {
-    use ironclaw_auth::CredentialAccountUpdateBinding;
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = test_service(filesystem, secret_store);
-
-    // Non-existent account in update_binding → CredentialMissing.
-    let err = service
-        .create_flow(NewAuthFlow {
-            scope: scope.clone(),
-            kind: AuthFlowKind::IntegrationCredential,
-            provider: google_provider(),
-            challenge: AuthChallenge::OAuthUrl {
-                authorization_url: OAuthAuthorizationUrl::new("https://provider.example/oauth")
-                    .unwrap(),
-                expires_at: Utc::now() + Duration::minutes(5),
-            },
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: Some(CredentialAccountUpdateBinding {
-                account_id: CredentialAccountId::new(),
-                ownership: CredentialOwnership::UserReusable,
-                owner_extension: None,
-                granted_extensions: vec![],
-            }),
-            opaque_state_hash: Some(state_hash("ubv-s")),
-            pkce_verifier_hash: Some(pkce_hash("ubv-p")),
-            expires_at: Utc::now() + Duration::minutes(5),
-        })
-        .await
-        .expect_err("non-existent binding account must return CredentialMissing");
-    assert_eq!(err, AuthProductError::CredentialMissing);
-}
-
-// ─── tests: update_status, project_credential_recovery, CredentialSetupService update ───
-
-#[tokio::test]
-async fn filesystem_update_status_and_cross_scope_rejection() {
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = test_service(filesystem, secret_store);
-
-    let account = service
-        .create_account(NewCredentialAccount {
-            scope: scope.clone(),
-            provider: google_provider(),
-            label: account_label(),
-            status: CredentialAccountStatus::Configured,
-            ownership: CredentialOwnership::UserReusable,
-            owner_extension: None,
-            granted_extensions: vec![],
-            access_secret: None,
-            refresh_secret: None,
-            scopes: vec![],
-        })
-        .await
-        .unwrap();
-
-    let updated = service
-        .update_status(&scope, account.id, CredentialAccountStatus::Inactive)
-        .await
-        .unwrap();
-    assert_eq!(updated.status, CredentialAccountStatus::Inactive);
-
-    // Non-existent account.
-    let err = service
-        .update_status(
-            &scope,
-            CredentialAccountId::new(),
-            CredentialAccountStatus::Inactive,
-        )
-        .await
-        .expect_err("missing account must return CredentialMissing");
-    assert_eq!(err, AuthProductError::CredentialMissing);
-}
-
-#[tokio::test]
-async fn filesystem_project_credential_recovery_returns_setup_required_when_empty() {
-    use ironclaw_auth::CredentialRecoveryRequest;
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = test_service(filesystem, secret_store);
-
-    // No accounts → setup_required.
-    let recovery = service
-        .project_credential_recovery(CredentialRecoveryRequest::new(
-            scope.clone(),
-            google_provider(),
-        ))
-        .await
-        .unwrap();
-    use ironclaw_auth::CredentialRecoveryState;
-    assert!(
-        matches!(
-            recovery.state,
-            CredentialRecoveryState::SetupRequired { .. }
-        ),
-        "no accounts must return setup_required"
-    );
-
-    // One configured account → configured.
-    let account = service
-        .create_account(NewCredentialAccount {
-            scope: scope.clone(),
-            provider: google_provider(),
-            label: account_label(),
-            status: CredentialAccountStatus::Configured,
-            ownership: CredentialOwnership::UserReusable,
-            owner_extension: None,
-            granted_extensions: vec![],
-            access_secret: None,
-            refresh_secret: None,
-            scopes: vec![],
-        })
-        .await
-        .unwrap();
-
-    let recovery = service
-        .project_credential_recovery(CredentialRecoveryRequest::new(
-            scope.clone(),
-            google_provider(),
-        ))
-        .await
-        .unwrap();
-    let CredentialRecoveryState::Configured { selected_account } = &recovery.state else {
-        panic!(
-            "single configured account must return Configured state, got: {:?}",
-            recovery.state
-        );
-    };
-    assert_eq!(selected_account.id, account.id);
-}
-
-#[tokio::test]
-async fn filesystem_credential_setup_service_update_path() {
-    use ironclaw_auth::{
-        CredentialAccountMutation, CredentialAccountUpdate, CredentialSetupService,
-    };
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = test_service(filesystem, secret_store);
-
-    let account = service
-        .create_account(NewCredentialAccount {
-            scope: scope.clone(),
-            provider: google_provider(),
-            label: account_label(),
-            status: CredentialAccountStatus::Configured,
-            ownership: CredentialOwnership::UserReusable,
-            owner_extension: None,
-            granted_extensions: vec![],
-            access_secret: Some(SecretHandle::new("old-access").unwrap()),
-            refresh_secret: None,
-            scopes: vec![],
-        })
-        .await
-        .unwrap();
-
-    let new_handle = SecretHandle::new("new-access").unwrap();
-    let updated = service
-        .create_or_update_account(CredentialAccountMutation::Update(CredentialAccountUpdate {
-            account_id: account.id,
-            account: NewCredentialAccount {
-                scope: scope.clone(),
-                provider: google_provider(),
-                label: account_label(),
-                status: CredentialAccountStatus::Configured,
-                ownership: CredentialOwnership::UserReusable,
-                owner_extension: None,
-                granted_extensions: vec![],
-                access_secret: Some(new_handle.clone()),
-                refresh_secret: None,
-                scopes: vec![],
-            },
-        }))
-        .await
-        .unwrap();
-    assert_eq!(updated.access_secret, Some(new_handle));
-}
-
-// ─── tests: get_account cross-scope rejection ─────────────────────────────────
-
-#[tokio::test]
-async fn filesystem_get_account_cross_scope_returns_cross_scope_denied() {
-    use ironclaw_auth::AuthSurface;
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = test_service(Arc::clone(&filesystem), Arc::clone(&secret_store));
-
-    let account = service
-        .create_account(NewCredentialAccount {
-            scope: scope.clone(),
-            provider: google_provider(),
-            label: account_label(),
-            status: CredentialAccountStatus::Configured,
-            ownership: CredentialOwnership::UserReusable,
-            owner_extension: None,
-            granted_extensions: vec![],
-            access_secret: None,
-            refresh_secret: None,
-            scopes: vec![],
-        })
-        .await
-        .unwrap();
-
-    // Same resource but different surface → CrossScopeDenied.
-    let cli_scope = AuthProductScope::new(scope.resource.clone(), AuthSurface::Cli);
-    let service2 = test_service(Arc::clone(&filesystem), Arc::clone(&secret_store));
-    let result = service2
-        .get_account(CredentialAccountLookupRequest::new(cli_scope, account.id))
-        .await;
-    // The account doesn't exist in the CLI path (different path on filesystem), so None.
-    assert!(
-        result.unwrap().is_none(),
-        "account written under web scope must not be visible under cli scope"
-    );
-}
-
-// ─── tests: validate_secret control-char branch ───────────────────────────────
-
-#[tokio::test]
-async fn filesystem_validate_secret_rejects_control_characters() {
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = test_service(filesystem, secret_store);
-
-    let challenge = service
-        .request_secret_input(ManualTokenSetupRequest {
-            scope: scope.clone(),
-            provider: google_provider(),
-            label: account_label(),
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: None,
-            expires_at: Utc::now() + Duration::minutes(5),
-        })
-        .await
-        .unwrap();
-    let AuthChallenge::ManualTokenRequired { interaction_id, .. } = challenge else {
-        panic!("expected ManualTokenRequired");
-    };
-
-    // NUL byte must be rejected without consuming the interaction.
-    let err = service
-        .submit_manual_token(
-            &scope,
-            SecretSubmitRequest {
-                interaction_id,
-                secret: SecretString::from("valid\x00nul"),
-            },
-        )
-        .await
-        .expect_err("NUL byte must be rejected");
-    assert!(
-        matches!(err, AuthProductError::InvalidRequest { .. }),
-        "must return InvalidRequest for control characters"
-    );
-
-    // Interaction must NOT be consumed — replay still possible.
-    let ok = service
-        .submit_manual_token(
-            &scope,
-            SecretSubmitRequest {
-                interaction_id,
-                secret: SecretString::from("clean-token"),
-            },
-        )
-        .await;
-    assert!(
-        ok.is_ok(),
-        "interaction must be usable after control-char rejection"
-    );
-}
-
-// ─── fix: abbyshekit review — expired flow mutation persisted ────────────────
-
-#[tokio::test]
-async fn filesystem_expired_flow_status_persisted_before_returning_error() {
-    // When claim_oauth_callback / complete_oauth_callback / fail_oauth_callback
-    // encounter an expired flow, the Expired status must be written to disk
-    // before returning UnknownOrExpiredFlow so durable state matches the contract.
-    use ironclaw_auth::{
-        AuthErrorCode, OAuthCallbackClaimRequest, OAuthCallbackFailureInput, OAuthCallbackInput,
-        ProviderCallbackOutcome,
-    };
-
-    let filesystem = test_filesystem();
-    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-    let scope = test_scope();
-    let service = test_service(filesystem, secret_store);
-
-    let flow = service
-        .create_flow(NewAuthFlow {
-            scope: scope.clone(),
-            kind: AuthFlowKind::IntegrationCredential,
-            provider: google_provider(),
-            challenge: AuthChallenge::OAuthUrl {
-                authorization_url: OAuthAuthorizationUrl::new("https://provider.example/oauth")
-                    .unwrap(),
-                expires_at: Utc::now() + Duration::minutes(5),
-            },
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: None,
-            opaque_state_hash: Some(state_hash("exp-s")),
-            pkce_verifier_hash: Some(pkce_hash("exp-p")),
-            expires_at: Utc::now() - Duration::seconds(1),
-        })
-        .await
-        .unwrap();
-
-    // claim_oauth_callback must persist Expired before returning error.
-    let err = service
-        .claim_oauth_callback(
-            &scope,
-            OAuthCallbackClaimRequest {
-                flow_id: flow.id,
-                opaque_state_hash: state_hash("exp-s"),
-                provider: google_provider(),
-                pkce_verifier_hash: pkce_hash("exp-p"),
-            },
-        )
-        .await
-        .expect_err("expired flow must be rejected");
-    assert_eq!(err, AuthProductError::UnknownOrExpiredFlow);
-
-    let persisted = service
-        .get_flow(&scope, flow.id)
-        .await
-        .unwrap()
-        .expect("flow must still exist");
-    assert_eq!(persisted.status, AuthFlowStatus::Expired);
-    assert_eq!(persisted.error, Some(AuthErrorCode::UnknownOrExpiredFlow));
-
-    // fail_oauth_callback on already-expired flow returns FlowAlreadyTerminal
-    // because Expired is a terminal status; the record was already persisted
-    // as Expired by claim_oauth_callback above.
-    let err2 = service
-        .fail_oauth_callback(
-            &scope,
-            OAuthCallbackFailureInput {
-                flow_id: flow.id,
-                opaque_state_hash: state_hash("exp-s"),
-                error: AuthErrorCode::ProviderDenied,
-            },
-        )
-        .await
-        .expect_err("expired flow must be rejected");
-    assert_eq!(
-        err2,
-        AuthProductError::FlowAlreadyTerminal,
-        "already-expired flow returns FlowAlreadyTerminal"
-    );
-
-    let persisted2 = service
-        .get_flow(&scope, flow.id)
-        .await
-        .unwrap()
-        .expect("flow must still exist");
-    assert_eq!(persisted2.status, AuthFlowStatus::Expired);
-    assert_eq!(persisted2.error, Some(AuthErrorCode::UnknownOrExpiredFlow));
-
-    // complete_oauth_callback on a fresh expired flow (never claimed) must also
-    // persist the Expired status before returning error.
-    let flow2 = service
-        .create_flow(NewAuthFlow {
-            scope: scope.clone(),
-            kind: AuthFlowKind::IntegrationCredential,
-            provider: google_provider(),
-            challenge: AuthChallenge::OAuthUrl {
-                authorization_url: OAuthAuthorizationUrl::new("https://provider.example/oauth")
-                    .unwrap(),
-                expires_at: Utc::now() + Duration::minutes(5),
-            },
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: None,
-            opaque_state_hash: Some(state_hash("exp2-s")),
-            pkce_verifier_hash: Some(pkce_hash("exp2-p")),
-            expires_at: Utc::now() - Duration::seconds(1),
-        })
-        .await
-        .unwrap();
-
-    let err3 = service
-        .complete_oauth_callback(
-            &scope,
-            OAuthCallbackInput {
-                flow_id: flow2.id,
-                opaque_state_hash: state_hash("exp2-s"),
-                outcome: ProviderCallbackOutcome::Denied,
-            },
-        )
-        .await
-        .expect_err("expired flow must be rejected");
-    assert_eq!(
-        err3,
-        AuthProductError::UnknownOrExpiredFlow,
-        "complete_oauth_callback on expired flow returns UnknownOrExpiredFlow"
-    );
-
-    let persisted3 = service
-        .get_flow(&scope, flow2.id)
-        .await
-        .unwrap()
-        .expect("flow2 must still exist");
-    assert_eq!(persisted3.status, AuthFlowStatus::Expired);
-    assert_eq!(persisted3.error, Some(AuthErrorCode::UnknownOrExpiredFlow));
-}
-
-// ─── fix: abbyshekit review — OAuth CAS-conflict branch purges old secrets ───
-
-#[tokio::test]
-async fn filesystem_oauth_cas_conflict_branch_purges_previous_secrets() {
-    // When the None-path CAS-conflict branch re-reads and overwrites an existing
-    // account, the previous access/refresh secret handles must be deleted from
-    // SecretStore so repeated re-auths do not accumulate dead handles.
-    use ironclaw_auth::ProviderCallbackOutcome;
-    use ironclaw_secrets::SecretMaterial;
-
-    let filesystem = test_filesystem();
-    let concrete_secret_store = Arc::new(InMemorySecretStore::new());
-    let secret_store: Arc<dyn SecretStore> = concrete_secret_store.clone();
-    let scope = test_scope();
-    let service = test_service(Arc::clone(&filesystem), Arc::clone(&secret_store));
-
-    let flow = service
-        .create_flow(NewAuthFlow {
-            scope: scope.clone(),
-            kind: AuthFlowKind::IntegrationCredential,
-            provider: google_provider(),
-            challenge: AuthChallenge::OAuthUrl {
-                authorization_url: OAuthAuthorizationUrl::new("https://provider.example/oauth")
-                    .unwrap(),
-                expires_at: Utc::now() + Duration::minutes(5),
-            },
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: None,
-            opaque_state_hash: Some(state_hash("cas-s")),
-            pkce_verifier_hash: Some(pkce_hash("cas-p")),
-            expires_at: Utc::now() + Duration::minutes(5),
-        })
-        .await
-        .unwrap();
-
-    // Pre-seed the account with old secrets.
-    let preseeded_id = CredentialAccountId::from_uuid(flow.id.as_uuid());
-    let old_access = SecretHandle::new("old-access").unwrap();
-    let old_refresh = SecretHandle::new("old-refresh").unwrap();
-    concrete_secret_store
-        .put(
-            scope.resource.clone(),
-            old_access.clone(),
-            SecretMaterial::from("old-access-token"),
-        )
-        .await
-        .unwrap();
-    concrete_secret_store
-        .put(
-            scope.resource.clone(),
-            old_refresh.clone(),
-            SecretMaterial::from("old-refresh-token"),
-        )
-        .await
-        .unwrap();
-
-    service
-        .create_account_with_id(
-            preseeded_id,
-            NewCredentialAccount {
-                scope: scope.clone(),
-                provider: google_provider(),
-                label: account_label(),
-                status: CredentialAccountStatus::Configured,
-                ownership: CredentialOwnership::UserReusable,
-                owner_extension: None,
-                granted_extensions: vec![],
-                access_secret: Some(old_access.clone()),
-                refresh_secret: Some(old_refresh.clone()),
-                scopes: vec![],
-            },
-            CasExpectation::Absent,
-        )
-        .await
-        .unwrap();
-
-    service
-        .claim_oauth_callback(
-            &scope,
-            OAuthCallbackClaimRequest {
-                flow_id: flow.id,
-                opaque_state_hash: state_hash("cas-s"),
-                provider: google_provider(),
-                pkce_verifier_hash: pkce_hash("cas-p"),
-            },
-        )
-        .await
-        .unwrap();
-
-    let new_access = SecretHandle::new("new-access").unwrap();
-    let new_refresh = SecretHandle::new("new-refresh").unwrap();
-    let completed = service
-        .complete_oauth_callback(
-            &scope,
-            OAuthCallbackInput {
-                flow_id: flow.id,
-                opaque_state_hash: state_hash("cas-s"),
-                outcome: ProviderCallbackOutcome::Authorized {
-                    exchange: OAuthProviderExchange {
-                        provider: google_provider(),
-                        account_label: account_label(),
-                        authorization_code_hash: code_hash("cas-c"),
-                        pkce_verifier_hash: pkce_hash("cas-p"),
-                        access_secret: new_access.clone(),
-                        refresh_secret: Some(new_refresh.clone()),
-                        scopes: vec![ProviderScope::new("gmail.readonly").unwrap()],
-                        account_id: None,
-                    },
-                },
-            },
-        )
-        .await
-        .unwrap();
-
-    assert_eq!(
-        completed.credential_account_id,
-        Some(preseeded_id),
-        "CAS-conflict branch must reuse pre-seeded account"
-    );
-
-    // Old secrets must be purged from SecretStore.
-    assert!(
-        concrete_secret_store
-            .metadata(&scope.resource, &old_access)
-            .await
-            .unwrap()
-            .is_none(),
-        "old access secret must be purged after CAS-conflict update"
-    );
-    assert!(
-        concrete_secret_store
-            .metadata(&scope.resource, &old_refresh)
-            .await
-            .unwrap()
-            .is_none(),
-        "old refresh secret must be purged after CAS-conflict update"
-    );
-}
-
-// ─── fix: abbyshekit review — manual-token consume only after success ────────
-
-#[tokio::test]
-async fn filesystem_manual_token_consume_only_after_successful_account_write() {
-    // If the account write fails, the interaction must NOT be marked consumed
-    // so the user can retry without going through a full re-setup.
-    use ironclaw_auth::CredentialAccountId;
-    use ironclaw_filesystem::CasExpectation;
-
-    let filesystem = test_filesystem();
-    let concrete_secret_store = Arc::new(InMemorySecretStore::new());
-    let secret_store: Arc<dyn SecretStore> = concrete_secret_store.clone();
-    let scope = test_scope();
-    let service = test_service(Arc::clone(&filesystem), Arc::clone(&secret_store));
-
-    let challenge = service
-        .request_secret_input(ManualTokenSetupRequest {
-            scope: scope.clone(),
-            provider: google_provider(),
-            label: account_label(),
-            continuation: AuthContinuationRef::SetupOnly,
-            update_binding: None,
-            expires_at: Utc::now() + Duration::minutes(5),
-        })
-        .await
-        .unwrap();
-    let AuthChallenge::ManualTokenRequired { interaction_id, .. } = challenge else {
-        panic!("expected ManualTokenRequired");
-    };
-
-    // Derive the account ID and pre-create a dummy record to force CAS failure.
-    let account_id = CredentialAccountId::from_uuid(interaction_id.as_uuid());
-    let dummy_account = ironclaw_auth::CredentialAccount {
-        id: account_id,
-        scope: scope.clone(),
-        provider: google_provider(),
-        label: account_label(),
-        status: CredentialAccountStatus::Configured,
-        ownership: CredentialOwnership::UserReusable,
-        owner_extension: None,
-        granted_extensions: vec![],
-        access_secret: None,
-        refresh_secret: None,
-        scopes: vec![],
-        created_at: Utc::now(),
-        updated_at: Utc::now(),
-    };
-    let path = super::paths::account_path(&scope, account_id)
-        .expect("account path derivation must succeed");
-    let json = serde_json::to_vec(&dummy_account).expect("serialization must succeed");
-    use ironclaw_filesystem::{ContentType, Entry};
-    let entry = Entry::bytes(json).with_content_type(ContentType::json());
-    filesystem
-        .put(&scope.resource, &path, entry, CasExpectation::Absent)
-        .await
-        .expect("pre-create dummy account must succeed");
-
-    // First submit fails because account write hits CAS conflict.
-    let err = service
-        .submit_manual_token(
-            &scope,
-            SecretSubmitRequest {
-                interaction_id,
-                secret: SecretString::from("first-attempt"),
-            },
-        )
-        .await
-        .expect_err("submit must fail when account write fails");
-    assert_eq!(
-        err,
-        AuthProductError::BackendConflict,
-        "CAS conflict must surface as BackendConflict"
-    );
-
-    // Interaction must NOT be consumed — retry still possible.
-    let retry_before_cleanup = service
-        .submit_manual_token(
-            &scope,
-            SecretSubmitRequest {
-                interaction_id,
-                secret: SecretString::from("retry-before-cleanup"),
-            },
-        )
-        .await;
-    assert!(
-        retry_before_cleanup.is_err(),
-        "retry must still fail because dummy account still blocks"
-    );
-    assert_eq!(
-        retry_before_cleanup.unwrap_err(),
-        AuthProductError::BackendConflict,
-        "retry must still hit BackendConflict, not UnknownOrExpiredFlow"
-    );
-
-    // Remove the dummy record so retry succeeds.
-    filesystem
-        .delete(&scope.resource, &path)
-        .await
-        .expect("delete dummy account must succeed");
-
-    let result = service
-        .submit_manual_token(
-            &scope,
-            SecretSubmitRequest {
-                interaction_id,
-                secret: SecretString::from("retry-token"),
-            },
-        )
-        .await;
-    assert!(
-        result.is_ok(),
-        "retry must succeed after removing the blocking dummy record"
-    );
-
-    // Third attempt must now fail with UnknownOrExpiredFlow because consumed_at is set.
-    let consumed_err = service
-        .submit_manual_token(
-            &scope,
-            SecretSubmitRequest {
-                interaction_id,
-                secret: SecretString::from("third-attempt"),
-            },
-        )
-        .await
-        .expect_err("third submit must fail because interaction is consumed");
-    assert_eq!(
-        consumed_err,
-        AuthProductError::UnknownOrExpiredFlow,
-        "interaction must be consumed after successful retry"
-    );
-}
-
-// ─── tests: broker projection contract (#4238) ───────────────────────────────
-//
-// These tests assert that every product-auth `write_account` call site
-// invokes the broker projector exactly once with the persisted record,
-// and that projector failures are non-fatal to the product-auth flow.
-
-mod broker_projection_tests {
-    use super::*;
-    use async_trait::async_trait;
-    use ironclaw_auth::{SecretCleanupAction, SecretCleanupRequest, SecretCleanupService};
-    use ironclaw_host_api::ExtensionId;
-    use std::sync::atomic::{AtomicUsize, Ordering};
-
-    /// Projector that fails on every call.  Used to assert product-auth
-    /// flows are resilient to broker projection failure.
-    struct FailingBrokerProjector {
-        calls: AtomicUsize,
-    }
-
-    impl FailingBrokerProjector {
-        fn new() -> Self {
-            Self {
-                calls: AtomicUsize::new(0),
-            }
-        }
-
-        fn calls(&self) -> usize {
-            self.calls.load(Ordering::SeqCst)
-        }
-    }
-
-    #[async_trait]
-    impl super::super::broker_projection::BrokerAccountProjector for FailingBrokerProjector {
-        async fn project_account(&self, _account: &ironclaw_auth::CredentialAccount) {
-            self.calls.fetch_add(1, Ordering::SeqCst);
-            // Real implementations would propagate a tracing::warn here.
-            // The trait contract is that this method MUST NOT panic and
-            // MUST NOT return an error.  We do nothing visible.
-        }
-    }
-
-    #[tokio::test]
-    async fn create_account_projects_configured_state_once() {
-        let filesystem = test_filesystem();
-        let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-        let scope = test_scope();
-        let (service, projector) = test_service_with_recording_projector(filesystem, secret_store);
-
-        let created = service
-            .create_account(NewCredentialAccount {
-                scope: scope.clone(),
-                provider: google_provider(),
-                label: account_label(),
-                status: CredentialAccountStatus::Configured,
-                ownership: CredentialOwnership::UserReusable,
-                owner_extension: None,
-                granted_extensions: vec![],
-                access_secret: Some(SecretHandle::new("oauth-access").unwrap()),
-                refresh_secret: None,
-                scopes: vec![],
-            })
-            .await
-            .unwrap();
-
-        let records = projector.records();
-        assert_eq!(records.len(), 1, "create_account must project exactly once",);
-        assert_eq!(records[0].id, created.id);
-        assert_eq!(records[0].status, CredentialAccountStatus::Configured);
-    }
-
-    #[tokio::test]
-    async fn cleanup_uninstall_projects_revoked_state() {
-        let filesystem = test_filesystem();
-        let concrete_secret_store = Arc::new(InMemorySecretStore::new());
-        let secret_store: Arc<dyn SecretStore> = concrete_secret_store.clone();
-        let scope = test_scope();
-        let (service, projector) = test_service_with_recording_projector(
-            Arc::clone(&filesystem),
-            Arc::clone(&secret_store),
-        );
-
-        let ext_id = ExtensionId::new("ext-projection").unwrap();
-        let access = SecretHandle::new("ext-access").unwrap();
-        use secrecy::SecretString;
-        concrete_secret_store
-            .put(
-                scope.resource.clone(),
-                access.clone(),
-                SecretString::from("material"),
-            )
-            .await
-            .unwrap();
-
-        let account = service
-            .create_account(NewCredentialAccount {
-                scope: scope.clone(),
-                provider: google_provider(),
-                label: account_label(),
-                status: CredentialAccountStatus::Configured,
-                ownership: CredentialOwnership::ExtensionOwned,
-                owner_extension: Some(ext_id.clone()),
-                granted_extensions: vec![],
-                access_secret: Some(access.clone()),
-                refresh_secret: None,
-                scopes: vec![],
-            })
-            .await
-            .unwrap();
-
-        service
-            .cleanup_for_lifecycle(SecretCleanupRequest {
-                scope: scope.clone(),
-                extension_id: ext_id.clone(),
-                action: SecretCleanupAction::Uninstall,
-            })
-            .await
-            .unwrap();
-
-        let records = projector.records();
-        assert_eq!(
-            records.len(),
-            2,
-            "create + cleanup must each emit one projection",
-        );
-        assert_eq!(records[0].status, CredentialAccountStatus::Configured);
-        assert_eq!(records[1].id, account.id);
-        assert_eq!(
-            records[1].status,
-            CredentialAccountStatus::Revoked,
-            "uninstall must project Revoked state to the broker",
-        );
-    }
-
-    #[tokio::test]
-    async fn projection_failure_does_not_block_product_auth_write() {
-        let filesystem = test_filesystem();
-        let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-        let scope = test_scope();
-        let projector = Arc::new(FailingBrokerProjector::new());
-        let service = FilesystemAuthProductServices::new(
-            filesystem,
-            secret_store,
-            Arc::clone(&projector)
-                as Arc<dyn super::super::broker_projection::BrokerAccountProjector>,
-        );
-
-        // The product-auth write must succeed even though the projector
-        // is configured to fail.  This is the "non-fatal but observable"
-        // contract from the broker_projection module docs.
-        let created = service
-            .create_account(NewCredentialAccount {
-                scope: scope.clone(),
-                provider: google_provider(),
-                label: account_label(),
-                status: CredentialAccountStatus::Configured,
-                ownership: CredentialOwnership::UserReusable,
-                owner_extension: None,
-                granted_extensions: vec![],
-                access_secret: None,
-                refresh_secret: None,
-                scopes: vec![],
-            })
-            .await
-            .expect("product-auth write must succeed despite projector failure");
-
-        // Read-back confirms durable record was persisted.
-        let fetched = service
-            .get_account(CredentialAccountLookupRequest::new(scope, created.id))
-            .await
-            .unwrap()
-            .expect("account must be durable");
-        assert_eq!(fetched.id, created.id);
-        assert_eq!(
-            projector.calls(),
-            1,
-            "projector must be invoked exactly once per write",
-        );
-    }
-
-    #[tokio::test]
-    async fn update_projects_each_write() {
-        use ironclaw_auth::CredentialAccountMutation;
-
-        let filesystem = test_filesystem();
-        let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
-        let scope = test_scope();
-        let (service, projector) = test_service_with_recording_projector(filesystem, secret_store);
-
-        let created = service
-            .create_account(NewCredentialAccount {
-                scope: scope.clone(),
-                provider: google_provider(),
-                label: account_label(),
-                status: CredentialAccountStatus::Configured,
-                ownership: CredentialOwnership::UserReusable,
-                owner_extension: None,
-                granted_extensions: vec![],
-                access_secret: None,
-                refresh_secret: None,
-                scopes: vec![],
-            })
-            .await
-            .unwrap();
-        assert_eq!(projector.records().len(), 1);
-
-        // Drive an update through CredentialSetupService.  Each write_account
-        // call (in flows, accounts, interactions, cleanup) must project.
-        use ironclaw_auth::{CredentialAccountUpdate, CredentialSetupService};
-        let update = CredentialAccountUpdate {
-            account_id: created.id,
-            account: NewCredentialAccount {
-                scope: scope.clone(),
-                provider: google_provider(),
-                label: CredentialAccountLabel::new("Alice Google (refreshed)").unwrap(),
-                status: CredentialAccountStatus::Configured,
-                ownership: CredentialOwnership::UserReusable,
-                owner_extension: None,
-                granted_extensions: vec![],
-                access_secret: Some(SecretHandle::new("rotated-access").unwrap()),
-                refresh_secret: None,
-                scopes: vec![],
-            },
-        };
-        let updated = service
-            .create_or_update_account(CredentialAccountMutation::Update(update))
-            .await
-            .unwrap();
-        assert_eq!(updated.id, created.id);
-
-        let records = projector.records();
-        assert_eq!(records.len(), 2, "create + update must each project once");
-        assert_eq!(records[1].id, created.id);
-        assert_eq!(
-            records[1].label.as_str(),
-            "Alice Google (refreshed)",
-            "projected record reflects updated label",
-        );
-    }
-}
