+++ crates/ironclaw_reborn_composition/src/product_auth_durable/tests/cleanup.rs
@@ -1,0 +1,198 @@
+use super::*;
+
+// ─── fix: durable SecretCleanupService purges secrets on Uninstall ───────────
+
+#[tokio::test]
+async fn filesystem_cleanup_for_lifecycle_deactivates_owner_and_revokes_on_uninstall() {
+    use ironclaw_auth::{SecretCleanupAction, SecretCleanupRequest, SecretCleanupService};
+    use ironclaw_host_api::ExtensionId;
+
+    let filesystem = test_filesystem();
+    let concrete_secret_store = Arc::new(InMemorySecretStore::new());
+    let secret_store: Arc<dyn SecretStore> = concrete_secret_store.clone();
+    let scope = test_scope();
+    let service = test_service(Arc::clone(&filesystem), Arc::clone(&secret_store));
+
+    let ext_id = ExtensionId::new("test-ext").unwrap();
+    let access = SecretHandle::new("ext-access").unwrap();
+    let refresh = SecretHandle::new("ext-refresh").unwrap();
+
+    // Seed secret material.
+    use secrecy::SecretString;
+    concrete_secret_store
+        .put(
+            scope.resource.clone(),
+            access.clone(),
+            SecretString::from("access-material"),
+        )
+        .await
+        .unwrap();
+    concrete_secret_store
+        .put(
+            scope.resource.clone(),
+            refresh.clone(),
+            SecretString::from("refresh-material"),
+        )
+        .await
+        .unwrap();
+
+    // Create an extension-owned account.
+    let account = service
+        .create_account(ironclaw_auth::NewCredentialAccount {
+            scope: scope.clone(),
+            provider: google_provider(),
+            label: account_label(),
+            status: CredentialAccountStatus::Configured,
+            ownership: CredentialOwnership::ExtensionOwned,
+            owner_extension: Some(ext_id.clone()),
+            granted_extensions: vec![],
+            access_secret: Some(access.clone()),
+            refresh_secret: Some(refresh.clone()),
+            scopes: vec![],
+        })
+        .await
+        .unwrap();
+
+    // Deactivate: account should be Inactive; secrets retained.
+    let deactivate_report = service
+        .cleanup_for_lifecycle(SecretCleanupRequest {
+            scope: scope.clone(),
+            extension_id: ext_id.clone(),
+            action: SecretCleanupAction::Deactivate,
+        })
+        .await
+        .unwrap();
+    assert_eq!(deactivate_report.retained_accounts, vec![account.id]);
+    assert!(
+        concrete_secret_store
+            .metadata(&scope.resource, &access)
+            .await
+            .unwrap()
+            .is_some(),
+        "Deactivate must retain secret material"
+    );
+
+    // Uninstall: account revoked, secrets purged from SecretStore.
+    let uninstall_report = service
+        .cleanup_for_lifecycle(SecretCleanupRequest {
+            scope: scope.clone(),
+            extension_id: ext_id.clone(),
+            action: SecretCleanupAction::Uninstall,
+        })
+        .await
+        .unwrap();
+    assert_eq!(uninstall_report.revoked_accounts, vec![account.id]);
+    assert!(
+        concrete_secret_store
+            .metadata(&scope.resource, &access)
+            .await
+            .unwrap()
+            .is_none(),
+        "Uninstall must delete access secret from SecretStore"
+    );
+    assert!(
+        concrete_secret_store
+            .metadata(&scope.resource, &refresh)
+            .await
+            .unwrap()
+            .is_none(),
+        "Uninstall must delete refresh secret from SecretStore"
+    );
+}
+
+// ─── fix: lock-cache weak-reference GC actually shrinks the map ──────────────
+
+#[tokio::test]
+async fn filesystem_lock_cache_drops_weak_entries_after_release() {
+    let filesystem = test_filesystem();
+    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
+    let service = test_service(filesystem, secret_store);
+
+    {
+        // Acquire a lock for key A and drop the guard immediately.
+        let lock_a = service.lock_for("account:key-a".to_string());
+        let _guard_a = lock_a.lock().await;
+        // guard_a dropped at end of this block; Arc<Mutex> dropped too after lock_a drops.
+    }
+    // After key-A's Arc dropped, the next call to lock_for should evict the
+    // dead weak reference. We trigger eviction via lock_for on a different key.
+    let _lock_b = service.lock_for("account:key-b".to_string());
+
+    // Verify key-A is gone: requesting it again must produce a *new* Arc (i.e.
+    // a fresh Mutex), not the evicted weak ref.
+    let lock_a2 = service.lock_for("account:key-a".to_string());
+    // The new lock should be unlocked (no one holds it).
+    assert!(
+        lock_a2.try_lock().is_ok(),
+        "re-acquired key-a must be unlocked"
+    );
+}
+
+// ─── fix: grant-removal on non-owner account in cleanup_for_lifecycle ─────────
+
+#[tokio::test]
+async fn filesystem_cleanup_removes_grant_from_non_owner_account() {
+    use ironclaw_auth::{SecretCleanupAction, SecretCleanupRequest, SecretCleanupService};
+    use ironclaw_host_api::ExtensionId;
+
+    let filesystem = test_filesystem();
+    let secret_store: Arc<dyn SecretStore> = Arc::new(InMemorySecretStore::new());
+    let scope = test_scope();
+    let service = test_service(filesystem, secret_store);
+
+    let ext_id = ExtensionId::new("granted-ext").unwrap();
+
+    // Create user-reusable account with a grant to ext_id (not owner).
+    let account = service
+        .create_account(NewCredentialAccount {
+            scope: scope.clone(),
+            provider: google_provider(),
+            label: account_label(),
+            status: CredentialAccountStatus::Configured,
+            ownership: CredentialOwnership::UserReusable,
+            owner_extension: None,
+            granted_extensions: vec![ext_id.clone()],
+            access_secret: None,
+            refresh_secret: None,
+            scopes: vec![],
+        })
+        .await
+        .unwrap();
+
+    let report = service
+        .cleanup_for_lifecycle(SecretCleanupRequest {
+            scope: scope.clone(),
+            extension_id: ext_id.clone(),
+            action: SecretCleanupAction::Uninstall,
+        })
+        .await
+        .unwrap();
+
+    assert_eq!(
+        report.removed_grants,
+        vec![account.id],
+        "grant must be reported removed"
+    );
+    assert!(
+        report.revoked_accounts.is_empty(),
+        "non-owner account must not be revoked"
+    );
+
+    let updated = service
+        .get_account(CredentialAccountLookupRequest::new(
+            scope.clone(),
+            account.id,
+        ))
+        .await
+        .unwrap()
+        .expect("account must still exist");
+    assert!(
+        !updated.granted_extensions.contains(&ext_id),
+        "grant must be removed from account record"
+    );
+    assert_eq!(
+        updated.status,
+        CredentialAccountStatus::Configured,
+        "status must be unchanged"
+    );
+}
